// src/app/impl/shared/model/AppCustBankStmntObjX.Model.ts
export class AppCustBankStmntObjX {
    AppCustBankStmntId: number;
    AppCustBankId: number;
    Month: string;
    Year: string;
    DebitAmt: number;
    DebitTrxCount: number;
    CreditAmt: number;
    CreditTrxCount: number;
    BalanceAmt: number;
    MonthName: string;

constructor() { 
        this.AppCustBankStmntId = 0;      
    }
}